engine: split tests
